+++ test/base/flop/validation_test.exs
@@ -26,13 +26,17 @@ defmodule Flop.ValidationTest do
 
   defmodule Thing do
     use Ecto.Schema
+    import Ecto.Query, only: [dynamic: 2]
 
     @derive {Flop.Schema,
              filterable: [],
-             sortable: [:name, :full_name, :thing_count],
+             sortable: [:name, :full_name, :thing_count, :thing_rank],
              adapter_opts: [
                compound_fields: [full_name: [:family_name, :given_name]],
-               alias_fields: [:thing_count]
+               alias_fields: [:thing_count],
+               custom_fields: [
+                 thing_rank: [sorter: {__MODULE__, :rank_sorter, []}]
+               ]
              ]}
 
     schema "things" do
@@ -40,6 +44,8 @@ defmodule Flop.ValidationTest do
       field :family_name, :string
       field :given_name, :string
     end
+
+    def rank_sorter(_opts), do: dynamic([t], t.name)
   end
 
   defp validate(params, opts \\ []) do
@@ -764,12 +770,18 @@ defmodule Flop.ValidationTest do
       assert {:error, changeset} = validate(params, for: Thing)
 
       assert errors_on(changeset)[:order_by] == [
-               "cursor pagination is not supported for compound and alias fields"
+               "cursor pagination is not supported for alias, compound and custom fields"
              ]
     end
 
     test "rejects an alias field as cursor order field" do
       params = %{first: 2, after: @cursor, order_by: [:thing_count]}
+      assert {:error, changeset} = validate(params, for: Thing)
+      assert errors_on(changeset)[:order_by] != nil
+    end
+
+    test "rejects a custom field as cursor order field" do
+      params = %{first: 2, after: @cursor, order_by: [:thing_rank]}
       assert {:error, changeset} = validate(params, for: Thing)
       assert errors_on(changeset)[:order_by] != nil
     end